requirements.txt.txt
Django==5.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyMuPDF==1.23.26
Pillow==10.1.0
gunicorn==21.2.0